Convert GeneFinder to tabbed components
